requirements.txt
pip install 

pyperclip
pyautogui
mysql-connector-python
schedule